refactor get_lastes_and_popular and getAllPogination

+++ src/main/java/com/example/airbnbb7/dto/response/ApplicationResponse.java
@@ -19,6 +19,5 @@ public class ApplicationResponse {
 
     private Long page;
 
-    private int pageSize;
-
+    private Long pageSize;
 }
